pyproject.toml
[tool.poetry]
name = "chiya"
version = "1.0.0"
description = "A moderation-heavy general purpose Discord bot."
license = "Unlicense"
authors = [
    "Snaacky <snaacky@pm.me>"
]
maintainers = [
    "Mint <developer@chocomint.dev>",
    "REDACTED <red4c13d@gmail.com>",
    "Kapppa <17926797+Kapppa@users.noreply.github.com>"
]
readme = "README.md"
repository = "https://github.com/Snaacky/Chiya"

[tool.poetry.dependencies]
python = "^3.11"
asyncpraw = "7.6.1"
dataset = "1.6.0"
"discord.py" = {version = "^2.1.0", extras = ["speed"]}
loguru = "^0.6.0"
mysqlclient = "2.1.1"
parsedatetime = "2.6"
PrivateBinAPI = "^1.0.0"
pyaml_env = "^1.2.1"
requests = "2.28.2"
sqlalchemy-utils = "0.39.0"

[tool.poetry.dev-dependencies]
black = ">=22.12.0"
flake8 = "^6.0.0"

[tool.black]
line-length = 120

[tool.poetry.extras]
"discord.py" = ["speed"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
